refactor(storage-ports): hoist scan-strategy surface from root (root-crate decomp)
Move the unified scan-strategy contract (ScanStrategy enum, ScanIterator /
UnifiedScanEngine traits, ScanStatistics, ScanCostEstimate, RowGroupStatistics,
scan_helpers) from root src/storage/scan_strategy.rs into the
proximadb-storage-ports crate alongside the other engine-port traits.
ScanCapabilities was already partially hoisted; this completes it.

The surface is a pure foundation-typed contract (proto Collection,
FilterExpression, ProximaRecord) every engine implements, so it belongs with
the port traits. Root src/storage/scan_strategy.rs becomes a one-line re-export
shim, so all crate::storage::scan_strategy::* paths (10 consumers across
traits / trait_components / engines) resolve unchanged.

This clears one of src/storage/traits' remaining real root-internal deps
(scan_strategy), continuing the Slice-D port-inversion chain toward
traits -> crate (link 7), which unblocks the 188K-symbol src/storage/engines
extraction (the jobs=2 / build-OOM lever).

Verified: root --lib + clippy --lib --bins clean; crate + crate-tests clippy
clean; fmt clean; storage up-edge ratchet unchanged at 197; work-commit-check
green (panic-policy net -1).
