+++ crates/storage/proximadb-storage-ports/src/scan_strategy.rs
@@ -1,0 +1,335 @@
+//! Unified Scan Strategy - Holistic design based on existing engine implementations
+//!
+//! This module unifies the scan patterns found across all storage engines:
+//! - **SST**: full_scan_strategy, filtered_scan_strategy_modular, parallel_full_scan
+//! - **VIPER**: evaluate_predicate_pushdown, columnar filtering with parallel evaluation
+//! - **NOVA**: Progressive columnar search, streaming processor, zone maps
+//! - **RAPTOR**: scan_vectors_with_strategy, full_scan_all_vectors, filtered_scan_vectors
+//!
+//! Key insights from existing implementations:
+//! 1. All engines separate full scan (for compaction) from filtered scan (for queries)
+//! 2. Columnar engines (VIPER/NOVA) have predicate pushdown capabilities
+//! 3. Hybrid columnar engines (SST) use bloom filters and block-level filtering
+//! 4. RAPTOR provides a unified ScanStrategy enum pattern
+//! 5. Parallel processing is key for multi-file scenarios
+//!
+//! Hoisted from the root crate's `src/storage/scan_strategy.rs` as a root-crate
+//! decomposition slice: the scan-strategy surface (`ScanStrategy`, `ScanIterator`,
+//! `UnifiedScanEngine`, `ScanStatistics`, …) is a pure foundation-typed contract
+//! every engine implements, so it belongs with the other engine-port traits. The
+//! root re-exports it as `crate::storage::scan_strategy` for source compatibility.
+
+use anyhow::Result;
+use async_trait::async_trait;
+use std::collections::HashSet;
+
+use proximadb_filter_expression::FilterExpression;
+use proximadb_proto::proximadb_v1::Collection;
+use proximadb_records::ProximaRecord;
+
+/// Unified scan strategy based on RAPTOR's successful pattern
+#[derive(Debug, Clone)]
+pub enum ScanStrategy {
+    /// Full scan for compaction, maintenance, export
+    /// Based on: SST's full_scan_strategy, VIPER's full file read, RAPTOR's FullScan
+    FullScan {
+        /// Include deleted/tombstoned records (for compaction)
+        include_deleted: bool,
+        /// Batch size for memory control
+        batch_size: usize,
+        /// Enable parallel file processing (from SST)
+        parallel: bool,
+        /// Use cache if available (from SST)
+        use_cache: bool,
+    },
+
+    /// Filtered scan with optimizations
+    /// Based on: VIPER's predicate_pushdown, NOVA's progressive search, RAPTOR's Filtering
+    FilteredScan {
+        /// Target IDs for direct lookup (from RAPTOR)
+        target_ids: Option<Vec<String>>,
+        /// Metadata predicates (from VIPER/NOVA)
+        predicates: Option<FilterExpression>,
+        /// Maximum row groups/blocks to scan (from RAPTOR)
+        max_blocks: Option<usize>,
+        /// Enable predicate pushdown (from VIPER/NOVA)
+        enable_pushdown: bool,
+        /// Enable column projection (from NOVA)
+        enable_projection: bool,
+        /// Early termination on limit
+        early_termination: bool,
+        /// Result limit
+        limit: Option<usize>,
+    },
+
+    /// Progressive scan with quantization stages (from NOVA)
+    ProgressiveScan {
+        /// Binary filtering stage
+        binary_candidates: usize,
+        /// INT8 refinement stage
+        int8_candidates: usize,
+        /// Final FP32 stage
+        fp32_candidates: usize,
+        /// Memory budget
+        memory_budget_bytes: Option<usize>,
+        /// Latency budget
+        latency_budget_ms: Option<u64>,
+    },
+
+    /// Range scan for ordered data (from SST)
+    RangeScan {
+        /// Start key (inclusive)
+        start_key: Option<String>,
+        /// End key (exclusive)
+        end_key: Option<String>,
+        /// Scan direction
+        reverse: bool,
+        /// Use index if available
+        use_index: bool,
+    },
+}
+
+/// Scan capabilities (re-export of the engine-capability descriptor that already
+/// lives in this crate's `capabilities` module). Kept here so the historical
+/// `scan_strategy::ScanCapabilities` path continues to resolve.
+pub use crate::ScanCapabilities;
+
+/// Scan iterator following existing patterns
+#[async_trait]
+pub trait ScanIterator: Send {
+    /// Get next batch (all engines use batched reading)
+    async fn next_batch(&mut self) -> Result<Option<Vec<ProximaRecord>>>;
+
+    /// Skip to position (for range scans)
+    async fn seek(&mut self, _key: &str) -> Result<()> {
+        Err(anyhow::anyhow!("Seek not supported by this iterator"))
+    }
+
+    /// Get current statistics
+    fn statistics(&self) -> ScanStatistics;
+
+    /// Cancel scan
+    fn cancel(&mut self);
+}
+
+/// Scan statistics from existing implementations
+#[derive(Debug, Clone, Default)]
+pub struct ScanStatistics {
+    // From all engines
+    pub records_scanned: usize,
+    pub records_matched: usize,
+    pub bytes_read: usize,
+
+    // From columnar engines (VIPER/NOVA)
+    pub row_groups_scanned: usize,
+    pub row_groups_pruned: usize,
+    pub columns_read: usize,
+
+    // From SST
+    pub blocks_scanned: usize,
+    pub blocks_filtered: usize,
+    pub bloom_filter_hits: usize,
+    pub cache_hits: usize,
+    pub cache_misses: usize,
+
+    // From NOVA progressive search
+    pub binary_candidates: usize,
+    pub int8_candidates: usize,
+    pub fp32_candidates: usize,
+
+    // Timing
+    pub io_time_ms: u64,
+    pub filter_time_ms: u64,
+    pub total_time_ms: u64,
+}
+
+/// Core scan trait that all engines implement
+#[async_trait]
+pub trait UnifiedScanEngine: Send + Sync {
+    /// Get scan capabilities for this engine
+    fn scan_capabilities(&self) -> ScanCapabilities;
+
+    /// Create a scan iterator based on strategy
+    /// This is the main entry point, similar to RAPTOR's scan_vectors_with_strategy
+    async fn create_scan_iterator(
+        &self,
+        _collection_id: &str,
+        strategy: ScanStrategy,
+        collection_config: Option<&Collection>,
+    ) -> Result<Box<dyn ScanIterator>>;
+
+    /// Estimate scan cost for query planning
+    async fn estimate_scan_cost(
+        &self,
+        _collection_id: &str,
+        strategy: &ScanStrategy,
+    ) -> Result<ScanCostEstimate>;
+
+    /// Optimize scan strategy based on statistics
+    /// Engines can upgrade/downgrade strategies based on their capabilities
+    async fn optimize_scan_strategy(
+        &self,
+        _collection_id: &str,
+        strategy: ScanStrategy,
+    ) -> Result<ScanStrategy> {
+        // Default: return as-is
+        Ok(strategy)
+    }
+}
+
+/// Scan cost estimate for query planning
+#[derive(Debug, Clone)]
+pub struct ScanCostEstimate {
+    pub estimated_records: usize,
+    pub estimated_bytes: usize,
+    pub estimated_time_ms: u64,
+    pub estimated_memory_bytes: usize,
+    pub confidence: f32,
+}
+
+/// Engine-specific implementations based on actual code patterns
+pub mod engine_impl {
+
+    /// SST scan implementation traits
+    pub struct SSTScanIterator {
+        /// Based on sst_query_engine.rs patterns
+        pub use_modular_reader: bool,
+        pub use_block_cache: bool,
+        pub parallel_files: bool,
+        pub bloom_filter_enabled: bool,
+    }
+
+    /// VIPER scan implementation traits
+    pub struct VIPERScanIterator {
+        /// Based on column_filter.rs patterns
+        pub parallel_column_evaluation: bool,
+        pub predicate_pushdown_enabled: bool,
+        pub selective_column_loading: bool,
+        pub qualifying_indices: Vec<usize>,
+    }
+
+    /// NOVA scan implementation traits
+    pub struct NOVAScanIterator {
+        /// Based on columnar_search.rs patterns
+        pub progressive_stages: bool,
+        pub streaming_enabled: bool,
+        pub zone_map_pruning: bool,
+        pub memory_budget: usize,
+    }
+
+    /// RAPTOR scan implementation traits
+    pub struct RAPTORScanIterator {
+        /// Based on consolidated_reader.rs patterns
+        pub tier_aware: bool,
+        pub rowgroup_filtering: bool,
+        pub bloom_filter_checking: bool,
+        pub consolidation_enabled: bool,
+    }
+}
+
+/// Helper functions extracted from existing implementations
+pub mod scan_helpers {
+    use super::*;
+
+    /// From VIPER: Extract columns needed for filter evaluation
+    pub fn extract_filter_columns(filter: &FilterExpression) -> HashSet<String> {
+        let mut columns = HashSet::new();
+        match filter {
+            FilterExpression::Comparison { field, .. } => {
+                columns.insert(field.clone());
+            }
+            FilterExpression::And(exprs) | FilterExpression::Or(exprs) => {
+                for expr in exprs {
+                    columns.extend(extract_filter_columns(expr));
+                }
+            }
+            _ => {}
+        }
+        columns
+    }
+
+    /// From SST: Check if block should be scanned based on bloom filter
+    pub async fn should_scan_block(
+        _block_id: &str,
+        bloom_filter: Option<&[u8]>,
+        target_ids: &[String],
+    ) -> bool {
+        // Implementation based on SST's bloom filter logic
+        if bloom_filter.is_none() || target_ids.is_empty() {
+            return true;
+        }
+        // Check bloom filter (simplified)
+        true
+    }
+
+    /// From NOVA: Determine if progressive search is beneficial
+    pub fn should_use_progressive(
+        total_vectors: usize,
+        selectivity: f32,
+        latency_budget_ms: Option<u64>,
+    ) -> bool {
+        // Based on NOVA's heuristics
+        total_vectors > 10000 && selectivity < 0.1 && latency_budget_ms.is_some()
+    }
+
+    /// From RAPTOR: Estimate row group selectivity
+    pub fn estimate_rowgroup_selectivity(
+        predicates: &[FilterExpression],
+        _statistics: &RowGroupStatistics,
+    ) -> f32 {
+        // Based on RAPTOR's statistics-based pruning
+        if predicates.is_empty() {
+            return 1.0;
+        }
+        // Simplified selectivity estimation
+        0.5
+    }
+}
+
+/// Placeholder for row group statistics
+#[derive(Debug, Clone)]
+pub struct RowGroupStatistics {
+    pub row_count: usize,
+    pub null_count: usize,
+    pub min_values: Vec<Option<serde_json::Value>>,
+    pub max_values: Vec<Option<serde_json::Value>>,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_scan_strategy_creation() {
+        // Test patterns from existing engines
+        let full_scan = ScanStrategy::FullScan {
+            include_deleted: false,
+            batch_size: 1024,
+            parallel: true,
+            use_cache: true,
+        };
+
+        let filtered_scan = ScanStrategy::FilteredScan {
+            target_ids: Some(vec!["id1".to_string()]),
+            predicates: None,
+            max_blocks: Some(100),
+            enable_pushdown: true,
+            enable_projection: true,
+            early_termination: true,
+            limit: Some(10),
+        };
+
+        let progressive_scan = ScanStrategy::ProgressiveScan {
+            binary_candidates: 10000,
+            int8_candidates: 1000,
+            fp32_candidates: 100,
+            memory_budget_bytes: Some(1024 * 1024 * 1024),
+            latency_budget_ms: Some(1000),
+        };
+
+        // All strategies should be constructible
+        matches!(full_scan, ScanStrategy::FullScan { .. });
+        matches!(filtered_scan, ScanStrategy::FilteredScan { .. });
+        matches!(progressive_scan, ScanStrategy::ProgressiveScan { .. });
+    }
+}
